% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{computation_time}
\alias{computation_time}
\title{Computation time (in hours) from MCMC output}
\usage{
computation_time(l)
}
\arguments{
\item{l}{an output from run_mcmc().}
}
\description{
Computation time (in hours) from MCMC output
}
