% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extraction}
\alias{extraction}
\title{Extract data from data/<date_str>.json & save as csv files}
\usage{
extraction(date_str)
}
\arguments{
\item{date_str}{a character string representing the date.}
}
\description{
Extract data from data/<date_str>.json & save as csv files
}
