% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{file_name}
\alias{file_name}
\title{Concatenate a date string & suffix}
\usage{
file_name(suffix, date_str)
}
\arguments{
\item{suffix}{a character string.}

\item{date_str}{a character string representing the date.}
}
\description{
Concatenate a date string & suffix
}
