% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{get_par_df}
\alias{get_par_df}
\title{Get and augment the parameter data frame of an MCMC output}
\usage{
get_par_df(df, fit_chr)
}
\arguments{
\item{df}{a data frame of the MCMC iterations of the parameters.}

\item{fit_chr}{a scalar character string for identifying the model/algorithm when combining with other outputs.}
}
\description{
Get and augment the parameter data frame of an MCMC output
}
