% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llik_nhpp_exp}
\alias{llik_nhpp_exp}
\title{Evaluate the log-likelihood under a special case of hybrid process when the power parameter is 0}
\usage{
llik_nhpp_exp(par, x)
}
\arguments{
\item{par}{vector of length 2, containing (in that order) the exponent parameter theta, and the scale parameter.}

\item{x}{vector of cumulative/absolute event times.}
}
\description{
Evaluate the log-likelihood under a special case of hybrid process when the power parameter is 0
}
