% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ms_etas}
\alias{ms_etas}
\title{Run the Gibbs variable selection algorithm for the hierarchical model of hybrid processes}
\usage{
ms_etas(beta, kappa, lambda, phi, psi, tau, theta, x, t_original, m_retweets,
  t_relative, t_infinity, s_beta, s_kappa, s_lambda, s_psi, s_theta, s_e_init,
  N, thin, burnin, print_freq, a_theta_prop, b_theta_prop, p0, p01, p10, write,
  filename_par = "default_par.csv", filename_m = "default_m.csv",
  filename_p = "default_p.csv", mu_beta = 0, tau_beta = 1e-04,
  mu_kappa = 0, tau_kappa = 1e-04, a_lambda = 1, b_lambda = 0.001,
  a_phi = 1, b_phi = 0.001, a_psi = 1, b_psi = 0.001, a_tau = 1,
  b_tau = 0.001, a_theta = 1, b_theta = 0.001)
}
\arguments{
\item{beta, kappa, lambda, phi, psi, tau, theta}{initial values of the parameters.}

\item{x, t_original, m_retweets, t_relative}{numeric vectors of follower counts, creation times of original tweets, number of retweets, and times of retweets relative to their corresponding original tweets.}

\item{t_infinity}{numeric scalar, duration of data collection.}

\item{s_beta, s_kappa, s_lambda, s_psi, s_theta, s_e_init}{initial values of proposal standard deviations.}

\item{N}{desired chain length AFTER burn-in and thinning.}

\item{thin}{thinning in MCMC.}

\item{burnin}{burn-in in MCMC.}

\item{print_freq}{how frequent should the current values in the MCMC be printed.}

\item{a_theta_prop}{shape parameter of pseudoprior for theta.}

\item{b_theta_prop}{rate parameter of pseudoprior for theta.}

\item{p0}{prior for power-law process as opposed to hybrid process.}

\item{p01}{jump probability from model 0 to 1 in RJMCMC.}

\item{p10}{jump probability from model 1 to 0 in RJMCMC.}

\item{write}{boolean; should csv files of estimated and predicted retweet counts be written?}

\item{filename_par, filename_m, filename_p}{names of the csv files for the chains of the parameters, the estimated retweet counts and the predicted retweet counts, respectively; ignored if write is FALSE.}

\item{mu_beta, tau_beta, mu_kappa, tau_kappa, a_lambda, b_lambda, a_phi, b_phi, a_psi, b_psi, a_tau, b_tau, a_theta, b_theta}{hyperparameters for the priors of beta, kappa, lambda, phi, psi, tau, and theta, respectively.}
}
\description{
Run the Gibbs variable selection algorithm for the hierarchical model of hybrid processes
}
