% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dpost}
\alias{plot_dpost}
\title{Overlay the posterior densities of MCMC output of different models/algorithms}
\usage{
plot_dpost(df, fit_str = c("0", "1"))
}
\arguments{
\item{df}{a data frame of the MCMC iterations.}

\item{fit_str}{a vector of character string, possibly defined in get_par_df().}
}
\description{
Overlay the posterior densities of MCMC output of different models/algorithms
}
