% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_retweet_count}
\alias{plot_retweet_count}
\title{Plot retweet count}
\usage{
plot_retweet_count(df, t_0, t_inf)
}
\arguments{
\item{df}{a data frame with at least the following variables: id_str and t_ij.}

\item{t_0}{start of data collection; POSIXct.}

\item{t_inf}{end of data collection relative to t_0; a positive scalar.}
}
\description{
Plot retweet count
}
