% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{read_data_csv}
\alias{read_data_csv}
\title{Read a csv file with specified date string & suffix}
\usage{
read_data_csv(suffix, date_str, ...)
}
\arguments{
\item{suffix}{a character string.}

\item{date_str}{a character string representing the date.}

\item{...}{additional arguments for the column types in readr::read_csv().}
}
\description{
Read a csv file with specified date string & suffix
}
