% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{reread}
\alias{reread}
\title{Re-read files if exist, extract if not}
\usage{
reread(date_str)
}
\arguments{
\item{date_str}{a character string representing the date.}
}
\description{
Re-read files if exist, extract if not
}
