% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Wrapper of the MCMC algorithms coded in C++}
\usage{
run_mcmc(date_str, seed, f_mcmc, N, thin, burnin, print_freq, a_theta_prop,
  b_theta_prop, p0, p01, p10, write, l0, beta, kappa, lambda, phi, psi, tau,
  theta, s_beta, s_kappa, s_lambda, s_psi, s_theta, s_e_init, rebuild = FALSE)
}
\arguments{
\item{date_str}{a character string representing the date.}

\item{seed}{a single value, interpreted as an integer, to be used in set.seed().}

\item{f_mcmc}{name of an MCMC algorithm coded in C++. Current choices are "mh_etas", "rj_etas" and "me_etas" (remove the quotes when using).}

\item{N}{desired chain length AFTER burn-in and thinning.}

\item{thin}{thinning in MCMC.}

\item{burnin}{burn-in in MCMC.}

\item{print_freq}{how frequent should the current values in the MCMC be printed.}

\item{a_theta_prop}{shape parameter of pseudoprior for theta.}

\item{b_theta_prop}{rate parameter of pseudoprior for theta.}

\item{p0}{prior for model 0 (power-law process), as opposed to model 1 (hybrid process).}

\item{p01}{jump probability from model 0 to 1 in RJMCMC.}

\item{p10}{jump probability from model 1 to 0 in RJMCMC.}

\item{write}{boolean; should csv files of estimated and predicted retweet counts be written?}

\item{l0}{a list containing two data frames df1.original and df0.range.}

\item{beta}{initial value of beta, coefficient for the linear term.}

\item{kappa}{initial value of kappa, coefficient for the quadratic term.}

\item{lambda}{initial value of lambda, parameter for the power-law effect in the hybrid process.}

\item{phi}{initial value of phi, scale parameter for the hybrid process.}

\item{psi}{initial value of psi, shift in the power-law term.}

\item{tau}{initial value of tau, precision of the Gaussian distribution for the random error.}

\item{theta}{initial value of theta, parameter for the exponential decay in the hybrid process.}

\item{s_beta}{initial proposal standard deviation for the Metropolis step for beta.}

\item{s_kappa}{initial proposal standard deviation for the Metropolis step for kappa.}

\item{s_lambda}{initial proposal standard deviation for the Metropolis step for lambda.}

\item{s_psi}{initial proposal standard deviation for the Metropolis step for psi.}

\item{s_theta}{initial proposal standard deviation for the Metropolis step for theta.}

\item{s_e_init}{initial proposal standard deviation for the Metropolis step for each random error.}

\item{rebuild}{boolean; should the MCMC be re-run if an output (an rds files) already exists?}
}
\description{
Wrapper of the MCMC algorithms coded in C++
}
