% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_and_mutate}
\alias{sim_and_mutate}
\title{Simulate hybrid process for multiple original tweets with respective follower counts and times of creation}
\usage{
sim_and_mutate(df.data, df.summary, t_inf)
}
\arguments{
\item{df.data}{a data frame with the following variables: id_str (character, unique identifier of original tweet), log_followers_count (logarithm of 1 + user follower count), t_i (time of creation of original tweet, relative to start of data collection).}

\item{df.summary}{a one-row data frame with the true values of the following parameters defined in sim_hybrid(): beta, kappa, lambda, phi, psi, tau, theta.}

\item{t_inf}{total duration of data collection / simulation period.}
}
\description{
Simulate hybrid process for multiple original tweets with respective follower counts and times of creation
}
