% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_hybrid}
\alias{sim_hybrid}
\title{Simulate retweet times according to hybrid process}
\usage{
sim_hybrid(T, x, beta, kappa, lambda, phi, psi, tau, theta)
}
\arguments{
\item{T}{duration of simulation period; a positive scalar.}

\item{x}{``mean-centred'' follower count; a real-valued scalar.}

\item{beta}{coefficient for the linear term x; a real-valued scalar.}

\item{kappa}{coefficient for the quadratic term x squared; a real-valued scalar.}

\item{lambda}{parameter for the power-law effect in the hybrid process; a scalar smaller than 1.}

\item{phi}{scale parameter for the hybrid process; a positive scalar.}

\item{psi}{shift in the power-law term; a non-negative scalar.}

\item{tau}{precision of the Gaussian distribution for the random error; a positive scalar.}

\item{theta}{parameter for the exponential decay in the hybrid process; a non-negative scalar.}
}
\description{
Simulate retweet times according to hybrid process
}
