% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_wrapper}
\alias{sim_wrapper}
\title{Wrapper of sim_and_mutate(), by creating a list that aligns with l.original}
\usage{
sim_wrapper(l.original, df.summary, seed)
}
\arguments{
\item{l.original}{a list containing two data frames df1.original and df0.range. The former can be used in the argument df.data in sim_and_mutate(), while the latter is a one-row data frame with (at least) the variable t_inf.}

\item{df.summary}{a one-row data frame with the true values of the following parameters defined in sim_hybrid(): beta, kappa, lambda, phi, psi, tau, theta.}

\item{seed}{a single value, interpreted as an integer, to be used in set.seed().}
}
\description{
Wrapper of sim_and_mutate(), by creating a list that aligns with l.original
}
